pyproject.toml
[tool.poetry]
package-mode = false
name = "tarantool-api"
version = "0.1.0"
description = ""
authors = ["Daniil Vilchinskiy <danvilcka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
fastapi = "^0.73.0"
uvicorn = "^0.14.0"
tarantool = "^0.6.3"
python-jose = "^3.2.0"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.14.0"